refactor: use dedicated service to manage Coin Data provider
