+++ src/containers/mobile/TokenDetailMobileContainer.tsx
@@ -35,8 +35,7 @@ import { getWeb3State } from "@/store/selectors";
 import { CHAIN_AVAILABLES } from "@/constants/chains";
 import { airplane, chevronDown, close, closeSharp, download, paperPlane } from "ionicons/icons";
 import { DataItem, SeriesData, SeriesMarkerData } from "@/components/ui/LightChart";
-import { getTokenHistoryPrice } from "@/utils/getTokenHistoryPrice";
-import { TokenInfo, getTokenInfo } from "@/utils/getTokenInfo";
+import { TokenInfo } from '@/servcies/coingecko.service';
 import { TokenDetailMarketDetail } from "@/components/ui/TokenDetailMarketData";
 import { TokenDetailDescription } from "@/components/ui/TokenDetailDescription";
 import { isStableAsset } from "@/utils/isStableAsset";
@@ -44,6 +43,7 @@ import { Currency } from "@/components/ui/Currency";
 import { NetworkTokenDetailCard } from "@/components/ui/NetworkTokenDetailCard/NetworkTokenDetailCard";
 import { getAllocationRatioInPercent } from "@/utils/getAllocationRatioInPercent";
 import { formatTxsAsSeriemarker } from "@/servcies/zerion.service";
+import { CoingeckoAPI } from "@/servcies/coingecko.service";
 
 const LightChart = lazy(() => import("@/components/ui/LightChart"));
 
@@ -79,10 +79,10 @@ export const TokenDetailMobileContainer = (props: {
     if (!walletAddress) return;
     const TxsSerie = formatTxsAsSeriemarker(txs, {symbol: data.symbol});
     setTxsChartHistory(()=> TxsSerie);
-    getTokenHistoryPrice(props.data.symbol).then((prices) => {
+    CoingeckoAPI.getTokenHistoryPrice(props.data.symbol).then((prices) => {
       setDataChartHistory(() => prices);
     });
-    getTokenInfo(props.data.symbol).then((tokenInfo) =>
+    CoingeckoAPI.getTokenInfo(props.data.symbol).then((tokenInfo) =>
       setTokenInfo(() => tokenInfo)
     );
   }, [walletAddress]);
